Remove notion of image name from objective-c processor
This made the macho objective-c processor aware of images and was passed around everywhere, instead we have an overridable section getter that gives the control over the section retrieval to the derived processor.

In the case of shared cache this means we can store the image to be processed and then use the header to locate the relevant sections.

Fixes: https://github.com/Vector35/binaryninja-api/issues/6594