Fix leaks from the Workflow constructor
Don't add an additional reference to the wrapped object in the
constructor.

API wrapper objects are passed a +1 object and do not need to add an
additional reference. All clients of the Workflow constructors are doing
this already which was resulting in leaks.

Fixes https://github.com/Vector35/binaryninja-api/issues/7059.