[CMake] Allow flexible user plugin install location
Create a cache variable of the same name as the previous normal variable
that users can set to a specific path for user plugin installation
destination.

This is primarily to help with CI builds to make it easier to find where
the final plugin is located before uploading as a release.